auto-download model in dit.New()
